handles QA generation w/o chapters

+++ app/backend/scraping.py
@@ -64,5 +64,6 @@ def read_file(file_name):
 
 
 def delete_file(file_name):
-    os.remove(file_name)
+    if (os.path.exists(file_name)):
+        os.remove(file_name)
     return
